debut du fix
l'endroit ou ca crash: le parser ne check pas bien si les choses sont null en cas de redir
