+++ parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: qjungo <qjungo@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 20:01:50 by qjungo            #+#    #+#             */
-/*   Updated: 2023/02/24 22:30:45 by qjungo           ###   ########.fr       */
+/*   Updated: 2023/03/09 14:26:05 by qjungo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,18 +65,22 @@ typedef struct s_cmd {
 typedef struct s_cmd		t_cmd;
 typedef struct s_minishell	t_minishell;
 
+// utils
 t_list		*skip_parentheses(t_list *cursor);
 int			get_token_type(char *token);
 t_bool		are_we_in_parentheses(t_list *tokens);
 t_list		*lst_cut_first_and_last(t_list *tokens);
-int			first_token_redir(t_list **tokens, t_cmd *cmd,
-				t_minishell *minishell);
 
+// parser
 int			parse_command(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 int			pipeline(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 int			set_command(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 t_cmd		*parser(t_list *tokens, t_minishell *minishell);
 
+// parser_redir
+int			redir(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
+
+// free_ast
 void		free_ast(t_cmd *cmd);
 
 int			is_heredoc_valid(t_list *cursor, t_cmd *cmd);
