+++ tokenizer/tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: qjungo <qjungo@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 20:17:58 by qjungo            #+#    #+#             */
-/*   Updated: 2023/03/07 11:18:57 by qjungo           ###   ########.fr       */
+/*   Updated: 2023/03/09 14:07:45 by qjungo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ int	tokenizer(const char *cmd, t_list **tokens, t_bool is_expand_mode,
 		cursor_index = token_position.end;
 		set_cursor_after_space(cmd, &cursor_index);
 	}
+	printf("|%s|> ", cmd);
+	LOG_TOKENS(*tokens);
 	return (SUCCESS);
 }
-//printf("|%s|> ", cmd);
-//LOG_TOKENS(*tokens);
